# Optimize icelake validate_utf32 for small inputs using masked AVX512 instructions