--- a/src/icelake/icelake_utf32_validation.inl.cpp
+++ b/src/icelake/icelake_utf32_validation.inl.cpp
@@ -18,7 +18,7 @@ bool validate_utf32(const char32_t *buf, size_t len) {
     currentmax = _mm512_max_epu32(utf32, currentmax);
   }
 
-  __m512i utf32 = _mm512_maskz_loadu_epi32(1 << (end - buf), buf);
+  __m512i utf32 = _mm512_maskz_loadu_epi32((__mmask16)1 << (end - buf), buf);
   currentoffsetmax =
       _mm512_max_epu32(_mm512_add_epi32(utf32, offset), currentoffsetmax);
   currentmax = _mm512_max_epu32(utf32, currentmax);