Modified the `validate_utf32` function to avoid using scalar loop for tail processing and to obtain the result using masked AVX512 instructions. Seems to speed up small inputs as observed below:
Before the changes:
```
validate_utf32+icelake, input size: 48, iterations: 1000, dataset: temp
   2.521 ins/byte,    0.542 cycle/byte,   10.139 GB/s (15.1 %),     5.492 GHz,    4.654 ins/cycle
  11.000 ins/char,    2.364 cycle/char,    2.323 Gc/s (15.1 %)     4.36 byte/char      4.7 ns
```

After the changes:
```
validate_utf32+icelake, input size: 48, iterations: 1000, dataset: temp
   1.333 ins/byte,    0.271 cycle/byte,   20.357 GB/s (1.8 %),     5.513 GHz,    4.923 ins/cycle
   5.818 ins/char,    1.182 cycle/char,    4.665 Gc/s (1.8 %)     4.36 byte/char      2.4 ns
```