@lemire 2024-12-28
Seems reasonable, merging.